Solution of SRAM bug in netparser: first trial
The SRAM bug is the one for which only 3 neurons can be connected from a source neuron.
This is not always the case. For example the connection of a neuron to a population of neurons in the same core can be modelled by writing only SRAM and the CAM of all destination neurons.