Solved Event Disappearing BUG and limitate connection BUG
BUG1 SOLVED: when you program three connection, the source neuron spikes disappears from the cAER spikeViewer.
BUG2 SOLVED: SRAM is always programmed for a connection, without looking if there are similar ones. This limits the fan-out of the neuron at 3. Now at least connections to the same chip and core are interpreted as similar, so SRAM is not programmed.