Added multi-core routing
NETPARSER module has been changed a bit to be able to exploit multi-core routing.
The changes has been made to let the network manager still store in the SRAM of the neuron every destination.
In this way it is makes very simple to retrieve all the connections from the network manager map.
This allows also to have a very simple network check: it is needed only to retrieve the connections from the network
and then compare with the original file.